+++ tests/issue-174.php
@@ -1,0 +1,10 @@
+<?php
+
+/**
+ * GitHub Issue:    https://github.com/ejmr/php-mode/issues/174
+ *
+ * Test highlighting of string literals containing escaped quotes
+ */
+
+'escaped single \' quotation mark';
+"escaped double \" quotation mark";