+++ tests/issue-197.php
@@ -1,0 +1,13 @@
+<?php
+
+/**
+ * GitHub Issue:    https://github.com/ejmr/php-mode/issues/197
+ *
+ * Test that member highlighting goes before highlighting of cc-mode
+ * types
+ */
+$test->array;
+$test->int;
+$test->string;
+$test->int();
+$test->string();