+++ tests/issue-201.php
@@ -1,0 +1,14 @@
+<?php
+
+/**
+ * GitHub Issue:    https://github.com/ejmr/php-mode/issues/201
+ *
+ * Test highighting of $this
+ */
+
+// Start:
+$this;
+$that;
+self::test();
+static::test();
+parent::test();